Updated vault table to use decimal numbers for d and e rating. Updated all references.

+++ RunApproachStatistics/Modules/Interfaces/IVaultModule.cs
@@ -36,8 +36,8 @@ namespace RunApproachStatistics.Modules.Interfaces
 
         List<int> getVaultKindIds();
 
-        List<vault> filter(double[] dRatings, double[] eRatings, int[] gymnasts, int[] locations, DateTime[] timestamps);
+        List<vault> filter(decimal[] dRatings, decimal[] eRatings, int[] gymnasts, int[] locations, DateTime[] timestamps);
 
-        List<vault> filter(double[] dRatings, double[] eRatings, String[] gymnasts, String[] locations, DateTime[] timestamps);
+        List<vault> filter(decimal[] dRatings, decimal[] eRatings, String[] gymnasts, String[] locations, DateTime[] timestamps);
     }
 }
